pyproject.toml
[tool.poetry]
name = "aws-dbt-pipeline"
version = "0.1.0"
description = ""
authors = ["oselio <oseliocandido@outlook.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12.2"
black = "^24.2.0"
pandas = "^2.2.1"
python-dotenv = "^1.0.1"
pre-commit = "^3.6.2"
mkdocs = "^1.5.3"
mkdocs-material = "^9.5.13"
streamlit = "^1.32.2"
pydantic = "^2.6.4"
openpyxl = "^3.1.2"
boto3 = "^1.34.64"
dbt-postgres = "^1.7.10"
colorama = "^0.4.6"
sqlalchemy = "^2.0.28"


[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"